Fix format specifier warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36566 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/widgets/ethernetcartwidget.c
@@ -72,7 +72,7 @@ static void on_base_changed(GtkWidget *widget, gpointer user_data)
     id = gtk_combo_box_get_active_id(GTK_COMBO_BOX(widget));
     base = (int)strtol(id, &endptr, 10);
     if (*endptr == '\0') {
-        debug_gtk3("setting ETHERNETCARTBase to $%04X.", base);
+        debug_gtk3("setting ETHERNETCARTBase to $%04X.", (unsigned int)base);
         resources_set_int("ETHERNETCARTBase", base);
     }
 }
